+++ media/gpu/mac/vt_config_util_unittest.cc
@@ -10,6 +10,7 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 #include "base/containers/span.h"
 #include "base/mac/foundation_util.h"
 #include "base/mac/mac_util.h"
+#include "base/numerics/safe_conversions.h"
 #include "base/strings/sys_string_conversions.h"
 #include "media/base/mac/color_space_util_mac.h"
 #include "media/formats/mp4/box_definitions.h"
@@ -44,7 +45,7 @@ base::span<const uint8_t> GetDataValue(CFDictionaryRef dict, CFStringRef key) {
       base::mac::CFCastStrict<CFDataRef>(CFDictionaryGetValue(dict, key));
   return data ? base::span<const uint8_t>(
                     reinterpret_cast<const uint8_t*>(CFDataGetBytePtr(data)),
-                    CFDataGetLength(data))
+                    base::checked_cast<size_t>(CFDataGetLength(data)))
               : base::span<const uint8_t>();
 }
